+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- Anthropic: Support for new `xhigh` value for `effort` (Claude 4.7).                       
+- Anthropic: Support for `max` value for `reasoning_effort`.   
+
 ## 0.3.206 (15 April 2026)
 
 - Eval Set: Display explicit `--id` in task panel headers when provided.